+++ tests/data/test_import_samples.py
@@ -50,18 +50,21 @@ class TestImportSamples(
         cls.sheet.cell(row=1, column=2, value="Id")
         cls.sheet.cell(row=1, column=3, value="Country")
         cls.sheet.cell(row=1, column=4, value="Sex")
+        cls.sheet.cell(row=1, column=5, value="Alias")
 
         # adding values
         cls.sheet.cell(row=2, column=1, value="TEX_IT")
         cls.sheet.cell(row=2, column=2, value="test-1")
         cls.sheet.cell(row=2, column=3, value="Italy")
         cls.sheet.cell(row=2, column=4, value="BHO")
+        cls.sheet.cell(row=2, column=5, value="test-one")
 
         # adding values
         cls.sheet.cell(row=3, column=1, value="TEX_IT")
         cls.sheet.cell(row=3, column=2, value="test-2")
         cls.sheet.cell(row=3, column=3, value="SPAIN")
         cls.sheet.cell(row=3, column=4, value="F")
+        cls.sheet.cell(row=3, column=5, value="test-two")
 
     def setUp(self):
         self.runner = CliRunner()
@@ -225,6 +228,62 @@ class TestImportSamples(
             # get the new sample: country will be always Italy
             sample = SampleSheep.objects.get(original_id="test-2")
             self.assertEqual(sample.smarter_id, "ITOA-TEX-000000002")
+
+    @patch('src.features.smarterdb.Dataset.working_dir',
+           new_callable=PropertyMock)
+    def test_import_alias(self, my_working_dir):
+        """Test importing samples with aliases"""
+
+        # create a temporary directory using the context manager
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            working_dir = pathlib.Path(tmpdirname)
+            my_working_dir.return_value = working_dir
+
+            # save worksheet in temporary folder
+            self.workbook.save(f"{working_dir}/metadata.xlsx")
+
+            # got first sample from database
+            self.assertEqual(SampleSheep.objects.count(), 1)
+
+            result = self.runner.invoke(
+                import_samples,
+                [
+                    "--src_dataset",
+                    "test2.zip",
+                    "--dst_dataset",
+                    "test.zip",
+                    "--datafile",
+                    "metadata.xlsx",
+                    "--code_column",
+                    "Code",
+                    "--country_column",
+                    "Country",
+                    "--id_column",
+                    "Id",
+                    "--alias_column",
+                    "Alias",
+                    "--chip_name",
+                    self.chip_name
+                ]
+            )
+
+            self.assertEqual(0, result.exit_code, msg=result.exception)
+
+            # I should have two record for samples, One already present one new
+            self.assertEqual(SampleSheep.objects.count(), 2)
+
+            # get the old sample
+            sample = SampleSheep.objects.get(original_id="test-1")
+            self.assertEqual(sample.smarter_id, "ITOA-TEX-000000001")
+
+            # import samples doesn't update attributes for existent samples
+            # HINT: should I update a sample?
+            self.assertIsNone(sample.alias)
+
+            # get the new sample
+            sample = SampleSheep.objects.get(original_id="test-2")
+            self.assertEqual(sample.smarter_id, "ESOA-TEX-000000002")
+            self.assertEqual(sample.alias, "test-two")
 
 
 if __name__ == '__main__':